Update auth logic
