+++ context/AuthContext/index.ts
@@ -1,0 +1,1 @@
+export { AuthContextProvider, useAuthContext } from "./AuthContext";
