+++ context/AuthContext/types.ts
@@ -1,0 +1,6 @@
+export interface AuthContextType {
+  isLoading: boolean;
+  loggedInUser: any | null;
+  refetchUserProfile: () => void;
+  signOut: () => void;
+}
